fix(configuration/git): address review comments +   split auth into sub-package
